DOPS-1820: refactor: split up plugin to be testable
